verify podcast catalogs (fix time and size from tags)
bit hacky but will expand on this later.

+++ src/Repository/Model/Podcast_Episode.php
@@ -558,25 +558,16 @@ class Podcast_Episode extends database_object implements Media, library_item, Ga
             $file    = $podcast->get_root_path();
             if (!empty($file)) {
                 $pinfo = pathinfo($this->source);
-
                 $file .= DIRECTORY_SEPARATOR . $this->pubdate . '-' . str_replace(array('?', '<', '>', '\\', '/'), '_', $this->title) . '-' . strtok($pinfo['basename'], '?');
-                debug_event(self::class, 'Downloading ' . $this->source . ' to ' . $file . ' ...', 4);
-                if (file_put_contents($file, fopen($this->source, 'r'))) {
-                    debug_event(self::class, 'Download completed.', 4);
-                    $this->file = $file;
-
-                    $vainfo = $this->getUtilityFactory()->createVaInfo($this->file);
-                    $vainfo->get_info();
-                    $key   = VaInfo::get_tag_type($vainfo->tags);
-                    $infos = VaInfo::clean_tag_info($vainfo->tags, $key, $file);
-                    // No time information, get it from file
-                    if ($this->time < 1) {
-                        $this->time = $infos['time'];
+                if (!is_file($file)) {
+                    debug_event(self::class, 'Downloading ' . $this->source . ' to ' . $file . ' ...', 4);
+                    if (file_put_contents($file, fopen($this->source, 'r'))) {
+                        debug_event(self::class, 'Download completed.', 4);
                     }
-                    $this->size = $infos['size'];
-
-                    $sql = "UPDATE `podcast_episode` SET `file` = ?, `size` = ?, `time` = ?, `state` = 'completed' WHERE `id` = ?";
-                    Dba::write($sql, array($this->file, $this->size, $this->time, $this->id));
+                }
+                if (is_file($file)) {
+                    debug_event(self::class, 'Updating details ' . $file . ' ...', 4);
+                    Catalog::update_media_from_tags($this);
                 } else {
                     debug_event(self::class, 'Error when downloading podcast episode.', 1);
                 }
